+++ CHANGELOG.md
@@ -6,13 +6,14 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 0.9.4
 
-2021 Feb 10 [in progress]
+2021 Feb 11 [in progress]
 
 *   Fixed Tools → Clear Accumulator #111
 *   Copied Single Value Widget value to clipboard on mouse click #113
 *   Used PySide.QtCore.Qt instead of PySide.QtGui.Qt #115
 *   Added error handling to log file cleanup.
 *   Fixed crash on invalid window_state #116
+*   Reduced exception catching from "except:" to "except Exception:"
 
 
 ## 0.9.3